x86/resctrl: Pass the schema in resdir's private pointer
We're trying to move bits that vary when using CDP out of the
rdt_resource structure into the new resctrl_schema, which can
eventually use the same resource in two different ways.

There are a few places where we really need the schema, but
only have the resdir's private pointer, which today is the
rdt_resource.

For the info dirs that represent a control, we need the schema,
as this is how the control is being used. For the monitors, we
need the resource as L3CODE_MON doesn't make sense, and would
monitor data too.

To do this, we need the types of the private pointers to vary
between control and monitor info dirs.

If the flags are RF_MON_INFO, its a struct rdt_resource. If the
flags are RF_CTRL_INFO, its a struct resctrl_schema. Nothing in
res_common_files[] has both flags.

Signed-off-by: James Morse <james.morse@arm.com>