x86/resctrl: Cull some uses of resctrl_to_arch_res()
Previously we littered resctrl_to_arch_res() wherever we needed to know the
conf_type of a cache. This pulled the value out of the arch specific
resource structure, which will eventually be removed.

Now that conf_type has a home, and we have the schema in all the right
places, fix all those callers to read the value from the relevant schema
entry.

Signed-off-by: James Morse <james.morse@arm.com>