x86/resctrl: Stop using Lx CODE/DATA resources
Now that CDP enable/disable is global, and the closid offset correction
is based on the configuration being applied, we are using different
hw_closid slots in the ctrl array for CODE/DATA schema. This lets
us merge them using the same Lx resource twice for CDP's CODE/DATA
schema. This keeps the illusion of separate caches in the resctrl code.

When CDP is enabled for a cache, create two schema generating the names
and setting the configuration type.

We can now remove the initialisation of of the illusionary hw_resources:
'cdp_capable' just requires setting a flag, resctrl knows what to do
from there.

Signed-off-by: James Morse <james.morse@arm.com>