x86/resctrl: Add a helper to avoid reaching into the arch code resource list
To hide arch specific details from resctrl we have the resctrl struct
embedded in an arch specific version. Resctrl currently pulls values
out of the architecture's arch-specific struct array.

Move the level enum to the resctrl header and add a helper to retrieve
the resctrl struct. resctrl_arch_get_resource() should not return
NULL for any value in the enum, it may instead return a dummy resource
that is !alloc_enabled && !mon_enabled.

Signed-off-by: James Morse <james.morse@arm.com>