x86/resctrl: Pull num_closid and num_rmid into arch helpers
Today resctrl exports a useless number to user space in the form
of a per-resource 'num_closids'. These can't all be used as
closid_init() rightly limits us to the system-wide minimum common
number of closids.

This gets exciting for Arm's MPAM, where num_closids may be
larger than resctrl supports, may vary wildly between resources,
and there may be resources that resctrl is unaware of, but must
still affect the system-wide minimum common number of closids.

Move the system-wide closid/rmid logic into arch helpers. Now resctrl
only ever knows about one value, we don't need to bounds-test the
closid when matching it with a resource and can remove the
per-resource values.

With this change resctrl needs to remember if it enabled CDP,
instead of twiddling the resource's num_closid.

We keep the control value arrays sized by hw_num_closid so that
components are reset with their default values, even if resctrl
is only using a subset of them.

The maximum closid resctrl can make use of is exposed as
RESCTRL_MAX_CLOSID in case the architecture supports significantly
more.

Signed-off-by: James Morse <james.morse@arm.com>