x86/resctrl: Switch closid setting/matchingto use helpers
When CDP is enabled Intel RDT uses closids in pairs, odd for code
even for data. We previously split the resctrl code up so that the
filesystem parts only worried about one closid, which with CDP had
two configurations. When calling the arch code to update the
configuration, we mapped the resctrl:closid to a hw_closid_t.

Add helpers for setting the closid for a task. Abstracting this will
allow another architecture to store different values in task_struct.

For MPAM, arm64 needs to store two closids and two rmids in task_struct,
one for each of code/data as there is no odd/even pairing.

Signed-off-by: James Morse <james.morse@arm.com>