x86/resctrl: Allow update_mba_bw() to update controls directly
There is no benefit to update_mba_bw() staging its config and cross
calling to finally commit the changes to hardware. We know we are on a
CPU associated with the domain, and this configuration doesn't come from
user-space. Instead it just modifies the msrs directly.

Abstract this behaviour into a new resctrl_arch_update_one() helper,
to live with the other arch-specific wrmsr helpers.

The delay_bw_map() moves to be part of the arch code, to maintain the
'percentage control' view of mba.

Signed-off-by: James Morse <james.morse@arm.com>