x86/resctrl: Abstract __rmid_read()
__rmid_read() reaches into the msrs of this CPU to pull out the value
of the specified event. Make the event an enum, and change the name
so that this is part of the arch specific api.

Excitingly, resctrl knows all about the hardware error bits in this
msr. Move the error handling into resctrl_arch_rmid_read(), mapping
RDT hardware errors onto errno values.

Signed-off-by: James Morse <james.morse@arm.com>