x86/resctrl: Convert use of rmid to rmid_idx
With RDT the rmid is an independent number used for counter events.
Each closid is allocated one or more. The relationship is arbitrary.

With MPAM, the pmg is effectively an extension to the partid space.
Each partid has the platforms number of pmg, whether it uses them or
not. For MPAM a monitor can only read a sensible value if its monitoring
a partid, and optionally a pmg.

To make this work we need resctrl_arch_rmid_read() to take a
partid/closid, and we need the partid/closid to be implicit when
updating the list of all monitor values.

For MPAM we also need resctrl to only allocate varying pmg bits to
control/monitor groups that share a closid/partid. But with RDT it
should be free to allocate all the rmids to one monitor group if
that's what the user wants.

To do this, we convert resctrl's monitor code to deal with an
rmid_idx. This is a value encoded by an arch helper. For RDT, its just
the rmid. With MPAM its the partid and pmg shift-and-masked together.

Store the hw_closid in the struct rmid_entry so it can be passed
to resctrl_arch_rmid_read(). This will later let us match the
closid/partid in the LRU list when allocating,

Signed-off-by: James Morse <james.morse@arm.com>