x86/resctrl: Pass clossid in to alloc/free rmid
Now that we always have the correct closid on hand when we allocate
an rmid, pass the value in to the alloc/free helpers. This means
that this value becomes part of the rmid_idx on architectures that
need to link these parameters.

x86 doesn't give you two rmids when using CDP, so we specify
CDP_BOTH for the hw_closid_t.

Signed-off-by: James Morse <james.morse@arm.com>