x86/resctrl: Pass closid in to resctrl_arch_rmid_read()
Finally closid is part of struct rmid_entry, and the two values can
be passed in to resctrl_arch_rmid_read(). This lets MPAM configure
its monitors, which require a partid as part of the match.

When iterating all monitor groups as part of the overflow thread, we need
to pass in the parent control group, so that the closid value can be
used.

Signed-off-by: James Morse <james.morse@arm.com>