arm64: cpufeature: discover CPU support for MPAM
ARMv8.4 adds support for 'Memory Partitioning And Monitoring' (MPAM)
which allows an interface to cache and bandwidth controls wherever
they appear in the system.

Add support to detect MPAM. MPAM's virtualisation support is optional
and detected separately so we can detect mismatched/insane systems, but
still use MPAM on the host even if the virtualisation support is
missing.

MPAM needs enabling at the highest implemented exception level. The
cpufeature 'matches' hook is extended to test this on one of the CPUs.

Signed-off-by: James Morse <james.morse@arm.com>
---
Testing that MPAM is enabled on all CPUs would requires MPAM1_EL1 being
described as an id register (which its not), so we can test the sanitised
enable bit. Instead we only test the enabled bit on the boot CPU.