arm64: mpam: Add arch hooks for MPAM user
Once MPAM has been succesfully probed by the driver and is in-use
we want to enable its context-switching code. Provide a static-key
and helpers to read/write it.

Driver code will also need to read the sanitised id registers to
find the common max partid/pmg values for all the online CPUs.

Signed-off-by: James Morse <james.morse@arm.com>

The user of this comes later, adding it early should save the arch
maintainer from spotting in the bigger mpam series.