cacheinfo: Provide a helper to find a cacheinfo leaf
To provide a reasonable threshold for a measure of cache-usage, MPAM
would like to keep track of the largest cache it has seen.

Provide a helper that takes a cpu and a cache-level, and returns
the corresponding leaf. This only works for unified caches.

Callers must hold the cpus_read_lock() to prevent the leaf being
free()d.

Signed-off-by: James Morse <james.morse@arm.com>