arm_mpam: Add stubs for cache psuedo lock / cache lockdown
We don't support resctrl's psuedo lock as the arm architecture already
has a way do to this using cache lockdown (D4.4.10 'Cache lockdown' of
DDI0487D.a). resctlr's approach can't be used in addition to this as
it relies on being able to prevent the CPU from accessing any other
regions of memory (e.g. the hardware prefetch) and to evict whatever
is in those cache 'portions' already.

Add stubs that return -EOPNOTSUPP for all the psuedo lock hooks.

Signed-off-by: James Morse <james.morse@arm.com>