DTS:DTS2019032609327 Description:crypto/hisilicon/qm: Fix two known issues
Team: HISI_SW
Feature or Bugfix: Bugfix

Original commit message as below:

This patch fixes below known issues:

1. When running 64 queues at same time, there are "eq overflow" error prints.
So we mask AEQ irq currently in this patch.

2. In test case 8, we set DEFAULT_THREAD_NUM = 256(requst 512 qps), it works
well. However, when we set DEFAULT_THREAD_NUM = 512(requst 1024 qps), all
tasks are timeout.

The reason of 1 is that we will trap in the loop of qm_irq_thread to poll
eqe, which means the updating of eq head may far behand real eq tail. This
will leads to eq overflow. So follow the suggestion in FS, we force to update
eq head when there are QM_Q_DEPTH / 2 - 1 EQEs.

The reason of 2 is there is a bug in qm_vft_data_cfg: if there is number
sq/cq, we should configure number - 1 to hardware. Stupid HW.

Change-Id: Ieec35645b9adc44a1c050f1c6df58737f37aba54
Signed-off-by: Zhou Wang <wangzhou1@hisilicon.com>
Reviewed-on: http://10.90.31.173:8080/5445
Tested-by: public TuringEE <turingee@huawei.com>
Reviewed-by: xuzibo 00494022 <xuzaibo@huawei.com>
Reviewed-by: public TuringEE <turingee@huawei.com>