x86/resctrl: Abstract and use supports_mba_mbps()
To determine whether the mba_mbps option to resctrl should be supported,
resctrl tests the boot cpus' x86_vendor. To make this portable, re-use
the tests set_mba_sc() does to determine if the mba_sc is supported
on this system.

An 'alloc_capable' test is added so that this property isn't implied
by 'linear'.

Signed-off-by: James Morse <james.morse@arm.com>