Documentation: x86: Clarify MBA takes MB as referring to mba_sc
"If the MBA is specified in MB then user can enter the max b/w in MB"
is a tautology. How can the user know if the schemata takes a percentage
or a MB/s value?

This is referring to whether the software controller is interpretting
the schemata's value. Make this clear.

Signed-off-by: James Morse <james.morse@arm.com>