arm_mpam: resctrl: Add resctrl and mpam configuration arrays
Now that we've picked the mpam classes to back the resctrl L2/L3
resources, we need somewhere to store the configuration.

We have two sets of configuration. The u32 from resctrl that it
expects to be a bitmap, and the values we generate from this to
write to the mpam devices.

Store both separately, the resctrl config lives in  mpam_resctrl_dom,
there is one array per resctrl resource. The mpam config lives in
struct mpam_component, which may be shared between multiple resctrl
resources.

This means mpam_devices.c gets one configuration for each mpam_component
even if its been supplied via two resctrl resources.

Sigend-off-by: James Morse <james.morse@arm.com>