arm_mpam: Add helpers to configure and read a monitor
If a component is made up of multiple devices, we need to configure
and read-from the monitor on each device. Extend what started as the
reset helper to configure and read from the monitors.

Reading from the monitor involves summing the value across multiple
devices. We could do better if we knew which devices had external
capture events, and how they were wired up. We don't, so we read the
counter value on each device.

Because this is currently only used for Cache Storage Usage we
aren't worried about overflow.

Signed-off-by: James Morse <james.morse@arm.com>