KVM: arm64: Make use of ARMv8.6 TWED feature
If supported and opted, use the TWED feature to add a delay in taking
a trap of WFE. The delay value(TWEDEL) is a 4-bit unsigned number and
encodes the minimum delay in taking a trap of WFE caused by
HCR_EL2.TWE as 2^(TWEDEL + 8) cycles.

The user can configure TWED enable and delay params. If enabled, a mid-range
value(0x7) is used as default TWEDEL.

Signed-off-by: Zengruan Ye <yezengruan@huawei.com>
Co-developed-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>
Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>