arm64/mm: Remove dependency on MM in new_context
The function new_context will be part of a generic ASID allocator. At
the moment, the MM structure is currently used to fetch the ASID and
pinned refcount.

To remove the dependency on MM, it is possible to just pass a pointer to
the current ASID and pinned refcount. Also please note that 'pinned' may
be NULL if the user doesn't require the pinned asid support.

v3-->v4:
  Changes related to Pinned ASID refcount.

Signed-off-by: Julien Grall <julien.grall@arm.com>
[Shameer: Changes related to Pinned ASID]
Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>