mlx5_vfio_pci: Implement vfio_pci driver for mlx5 devices
This patch adds support for vfio_pci driver for mlx5 devices.

It uses vfio_pci_core to register to the VFIO subsystem and then
implements the mlx5 specific logic in the migration area.

The migration implementation follows the definition from uapi/vfio.h and
uses the mlx5 VF->PF command channel to achieve it.

This patch implements the suspend/resume flows.

issue: 2690413
Change-Id: I0a0c63e418f3f4ee10b55c676e66307d6adb2446
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>