PCI: Add a PCI_ID_F_VFIO_DRIVER_OVERRIDE flag to struct pci_device_id
The new flag field is be used to allow PCI drivers to signal the core code
during driver matching and when generating the modules.alias information.

The first use will be to define a VFIO flag that indicates the PCI driver
is a VFIO driver.

VFIO drivers have a few special properties compared to normal PCI drivers:
 - They do not automatically bind. VFIO drivers are used to swap out the
   normal driver for a device and convert the PCI device to the VFIO
   subsystem.

   The admin must make this choice and following the current uAPI this is
   usually done by using the driver_override sysfs.

 - The modules.alias includes the IDs of the VFIO PCI drivers, prefixing
   them with 'vfio_pci:' instead of the normal 'pci:'.

   This allows the userspace machinery that switches devices to VFIO to
   know what kernel drivers support what devices and allows it to trigger
   the proper device_override.

As existing tools do not recognize the "vfio_pci:" mod-alias prefix this
keeps todays behavior the same. VFIO remains on the side, is never
autoloaded and can only be activated by direct admin action.

This patch is the infrastructure to provide the information in the
modules.alias to userspace and enable the only PCI VFIO driver. Later
series introduce additional HW specific VFIO PCI drivers.

Userspace can switch a device to the VFIO flavour of driver by using
something like this algorithm:

   1) Identify the sysfs path to the device:
    /sys/devices/pci0000:00/0000:00:01.0/0000:01:00.0

   2) Get the modalias file:
    $ cat /sys/bus/pci/devices/0000:01:00.0/modalias
   pci:v000015B3d00001017sv000015B3sd00000001bc02sc00i00

   3) Prefix it with vfio_:
   vfio_pci:v000015B3d00001017sv000015B3sd00000001bc02sc00i00

   4) Search modules.alias for the above string, select the entry that has
      the fewest *'s.

   5) modprobe the matched module name

   6) cat the matched modules.alias module name to
      /sys/bus/pci/devices/0000:01:00.0/driver_override

issue: 2690413
Change-Id: Ia48d440a0203f3b93dbed4f20025e34b5d31f349
Signed-off-by: Max Gurtovoy <mgurtovoy@nvidia.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>