vfio/hisilicon: add acc live migration driver
This driver adds the code required by Hisilicon
accelerator device to realize the live migration function.
It mainly includes the following functions:
(1).Match the accelerator device with the vfio-pci
driver framework.
(2).Processing of the status of the live migration
function and processing of the migration data.
(3).Operation and data processing of accelerator hardware devices

Signed-off-by: Longfang Liu <liulongfang@huawei.com>
[Shameer: Changes related to new vfio-pci-core framework]
Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>