net/mlx5: Introduce migration bits and structures
Introduce migration IFC related stuff to enable migration commands as of
below.

The SUSPEND command should be used to quiesce / freeze the device.
The RESUME command should be used to un-quiesce / un-freeze the device.
The QUERY STATE command should be used to read the migration state size.
The SAVE STATE command should be used to read from the device its state.
The LOAD STATE command should be used to write to the device state data.

issue: 2690413
Change-Id: Ic600bdf43a8b1efda8bbd63b3ee7fec1176ed95a
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>